First Commit
Fizz buzz is a group word game for children to teach them about division. Players take turns to count incrementally, replacing any number divisible by three with the word "fizz", and any number divisible by five with the word "buzz". This web app tells the user the amount of Fizzes and Buzzes are in a number. 

Two things to be added:
1) Better CSS; Better looking site
2) Option for user to "guess" then reveal the answer